requirements.txt
Flask==2.3.0
flask-cors==4.0.0
python-dotenv==1.0.0
razorpay==1.3.0
Pillow>=10.2.0
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0
qrcode==7.4.2
elevenlabs==2.39.0
